+++ examples/azure/snapshots-copy/config.sh
@@ -1,0 +1,8 @@
+#!/usr/bin/env nash
+
+## Resource Group Settings
+group       = "klb-examples-snapshots-copy"
+location    = "eastus2"
+sku         = "Standard_LRS"
+disk_size   = "10"
+other_location = "westus"