+++ CHANGELOG.md
@@ -1,3 +1,24 @@
+# Release 0.11.0
+
+This release breaks compatibility of the vm package.
+
+The functions that have changed are:
+
+* azure_vm_backup_create
+* azure_snapshot_create
+
+To be more compatible with azure behavior of not allowing snapshots to
+be on a different location than the source disks used to generate them.
+
+We also added new functionality to enable backups to be copied through
+different locations, check the functions:
+
+* azure_vm_backup_copy
+* azure_snapshot_copy
+
+There are examples provided on the project for both functionalities.
+
+
 # Release 0.10.0
 
 This release breaks compatibility of the storage package to